[xcc] fixed bug in amo{maxu,minu}.w

+++ riscv/insns/amominu_w.h
@@ -1,3 +1,3 @@
 uint32_t v = mmu.load_int32(RS1);
 mmu.store_uint32(RS1, std::min(uint32_t(RS2),v));
-RD = v;
+RD = (int32_t)v;
